Extractor/requirements.txt
aiofiles==24.1.0
aiohttp==3.11.18
beautifulsoup4==4.13.4
crawl4ai==0.6.3
Requests==2.32.3
PyYAML==6.0.1
fastapi
uvicorn[standard]
feedparser